requirements.txt
discord.py
python-dotenv
beautifulsoup4
httpx